+++ packages/side-effects/side-effect-title/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 1.0.0-beta.9
 
 * Side effect receives `Emitted`.